+++ src/mol-theme/color/volume-value.ts
@@ -1,17 +1,17 @@
 /**
- * Copyright (c) 2021-2022 mol* contributors, licensed under MIT, See LICENSE file for more info.
+ * Copyright (c) 2021-2025 mol* contributors, licensed under MIT, See LICENSE file for more info.
  *
  * @author Alexander Rose <alexander.rose@weirdbyte.de>
  */
 
-import type { ColorTheme } from '../color';
-import { Color, ColorScale } from '../../mol-util/color';
+import { ColorTheme } from '../color';
 import { ParamDefinition as PD } from '../../mol-util/param-definition';
 import { ThemeDataContext } from '../theme';
 import { ColorNames } from '../../mol-util/color/names';
 import { ColorTypeDirect } from '../../mol-geo/geometry/color-data';
 import { Volume } from '../../mol-model/volume/volume';
 import { ColorThemeCategory } from './categories';
+import { normalize } from '../../mol-math/interpolate';
 
 const Description = 'Assign color based on the given value of a volume cell.';
 
@@ -26,6 +26,12 @@ export const VolumeValueColorThemeParams = {
             [ColorNames.white, 1]
         ]
     }, { offsets: true, isEssential: true }),
+    domain: PD.MappedStatic('auto', {
+        custom: PD.Interval([-1, 1], { step: 0.001 }),
+        auto: PD.Group({
+            symmetric: PD.Boolean(false, { description: 'If true the automatic range is determined as [-|max|, |max|].' })
+        })
+    }),
 };
 export type VolumeValueColorThemeParams = typeof VolumeValueColorThemeParams
 export function getVolumeValueColorThemeParams(ctx: ThemeDataContext) {
@@ -33,21 +39,32 @@ export function getVolumeValueColorThemeParams(ctx: ThemeDataContext) {
 }
 
 export function VolumeValueColorTheme(ctx: ThemeDataContext, props: PD.Values<VolumeValueColorThemeParams>): ColorTheme<VolumeValueColorThemeParams, ColorTypeDirect> {
-    const scale = ColorScale.create({ domain: [0, 1], listOrName: props.colorList.colors });
+    let palette: ColorTheme.Palette | undefined;
 
-    const colors: Color[] = [];
-    for (let i = 0; i < 256; ++i) {
-        colors[i] = scale.color(i / 255);
+    if (ctx.volume) {
+        const { min, max } = ctx.volume.grid.stats;
+        const domain: [number, number] = props.domain.name === 'custom' ? props.domain.params : [min, max];
+        const { colorList } = props;
+
+        if (props.domain.name === 'auto' && props.domain.params.symmetric) {
+            const max = Math.max(Math.abs(domain[0]), Math.abs(domain[1]));
+            domain[0] = -max;
+            domain[1] = max;
+        }
+
+        const normalizedDomain = [
+            normalize(domain[0], min, max),
+            normalize(domain[1], min, max)
+        ] as [number, number];
+
+        palette = ColorTheme.Palette(colorList.colors, colorList.kind, normalizedDomain);
     }
-
-    const palette: ColorTheme.Palette = { colors, filter: 'linear' };
 
     return {
         factory: VolumeValueColorTheme,
         granularity: 'direct',
-        props: props,
+        props,
         description: Description,
-        legend: scale.legend,
         palette,
     };
 }