+++ CHANGELOG.md
@@ -16,6 +16,9 @@ Note that since we don't clearly distinguish between a public and private interf
 - Add `external-structure` theme that colors any geometry by structure properties
 - Support float and half-float data type for direct-volume rendering and GPU isosurface extraction
 - Minor documentation updates
+- Add `volume-data` theme that colors positions by volume data
+- Add support for color themes to `slice` representation
+- Improve/fix palette support in volume color themes
 
 ## [v4.10.0] - 2024-12-15